Updated edgetree.h exported for the LinkingTo functions to incrporate changes to that in src/.
git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/branches/3.1@12528 9cb281f4-553f-406f-8d7f-11f072bbfe46

+++ inst/include/edgetree.h
@@ -1,4 +1,4 @@
-/*  File inst/include/edgetree.h in package ergm, part of the Statnet suite
+/*  File src/edgetree.h in package ergm, part of the Statnet suite
  *  of packages for network analysis, http://statnet.org .
  *
  *  This software is distributed under the GPL-3 license.  It is free,
@@ -17,7 +17,7 @@
 
 #define MIN(a,b) ((a)<(b) ? (a) : (b))
 #define MAX(a,b) ((a)<(b) ? (b) : (a))
-#define DYADCOUNT(nnodes, bipartite, directed) (bipartite? (nnodes-bipartite)*bipartite : (directed? nnodes*(nnodes-1) : (nnodes*(nnodes-1))/2))
+#define DYADCOUNT(nnodes, bipartite, directed) (bipartite? (unsigned long)(nnodes-bipartite)*(unsigned long)bipartite : (directed? (unsigned long)nnodes*(unsigned long)(nnodes-1) : (((unsigned long)nnodes*(unsigned long)(nnodes-1))/2)))
 
 /*typedef unsigned int Vertex;
 typedef unsigned int Edge; */
